[feat/#17] 지역 컨트롤러 및 서비스 구현

+++ src/main/java/com/app/weather/domain/region/domain/Region.java
@@ -24,9 +24,9 @@ public class Region extends BaseEntity {
 
     private String name;
 
-    private int lat;
+    private double lat;
 
-    private int lon;
+    private double lon;
 
     @OneToMany(mappedBy = "region",cascade = CascadeType.PERSIST)
     @Builder.Default
